requirements-docker.txt
# pyannote-audio 4.0.4 — supports community-1 pipeline
# Its torch==2.8.0 pin matches our pytorch/pytorch:2.8.0 base image
pyannote-audio==4.0.4

# torchcodec 0.7.0 — audio decoding for transformers ASR pipeline
# MUST match PyTorch version: 0.6/0.7 for torch 2.8 (see compatibility table)
# https://github.com/pytorch/torchcodec?tab=readme-ov-file#installing-torchcodec
torchcodec==0.7.0

# ASR
transformers>=4.48.0
accelerate>=0.27.2

# Server
fastapi[standard]
uvicorn

# HTTP client (for audio_url downloads)
httpx

# Config
pydantic>=2.6.3
pydantic-settings>=2.2.1
